pixelharness: fix MSVC warning about intptr_t conversion to int

+++ source/test/pixelharness.cpp
@@ -619,7 +619,7 @@ bool PixelHarness::check_cvt16to32_cnt_t(cvt16to32_cnt_t ref, cvt16to32_cnt_t op
         memset(ref_dest, 0xCD, sizeof(ref_dest));
         memset(opt_dest, 0xCD, sizeof(opt_dest));
 #endif
-        int opt_cnt = checked(opt, opt_dest, sbuf1 + j, stride);
+        int opt_cnt = (int)checked(opt, opt_dest, sbuf1 + j, stride);
         int ref_cnt = ref(ref_dest, sbuf1 + j, stride);
 
         if ((ref_cnt != opt_cnt) || memcmp(ref_dest, opt_dest, 64 * 64 * sizeof(int32_t)))
